requirements.txt
Flask
Werkzeug
requests
dnspython
phonenumbers
Pillow
PyPDF2
python-docx
duckduckgo-search
groq
beautifulsoup4
lxml
urllib3